Flatten the atmo package a bit
This gets rid of atmo.utils in favor of a flatter package layout.
Renames the atmo.schedule module to atmo.jobs to prevent overlap with atmo.scheduling.
