+++ atmo/clusters/models.py
@@ -5,7 +5,7 @@ from django.db import models
 from django.contrib.auth.models import User
 from django.utils import timezone
 
-from ..utils import provisioning
+from .. import provisioning
 
 
 class Cluster(models.Model):
